requirements.txt
langchain
langchain-openai
python-dotenv
#beautifulsoup4
#chromadb
python-box